+++ src/model/impls/Grade.java
@@ -1,4 +1,4 @@
-package model;
+package model.impls;
 
 public enum Grade {
     A, B, C, D, F, N
