+++ src/model/impls/GradeCalculator.java
@@ -1,4 +1,4 @@
-package model;
+package model.impls;
 
 public class GradeCalculator {
     private static final int[] mandatoryThresholds = {95, 90, 80, 70, 60, 0};
