+++ src/model/impls/Status.java
@@ -1,4 +1,4 @@
-package model;
+package model.impls;
 
 public enum Status {
     GREEN,
